Componentize config screen
